Fixed tests for soft dirty PTEs
