+++ tests/test/scheduler/test_executor.cpp
@@ -54,9 +54,9 @@ void TestExecutor::restore(faabric::Message& msg)
     snap->mapToMemory({ dummyMemory.get(), dummyMemorySize });
 }
 
-faabric::util::MemoryView TestExecutor::getMemoryView()
+std::span<uint8_t> TestExecutor::getMemoryView()
 {
-    return faabric::util::MemoryView({ dummyMemory.get(), dummyMemorySize });
+    return { dummyMemory.get(), dummyMemorySize };
 }
 
 int32_t TestExecutor::executeTask(
